+++ cisstMultiTask/code/mtsManagerLocal.cpp
@@ -5,7 +5,7 @@
   Author(s):  Min Yang Jung
   Created on: 2009-12-07
 
-  (C) Copyright 2009-2020 Johns Hopkins University (JHU), All Rights Reserved.
+  (C) Copyright 2009-2024 Johns Hopkins University (JHU), All Rights Reserved.
 
 --- begin cisst license - do not edit ---
 
@@ -20,6 +20,7 @@ http://www.cisst.org/cisst/license.txt.
 
 #include <cisstCommon/cmnThrow.h>
 #include <cisstCommon/cmnPath.h>
+#include <cisstCommon/cmnPortability.h>
 #include <cisstOSAbstraction/osaSleep.h>
 #include <cisstOSAbstraction/osaGetTime.h>
 #include <cisstOSAbstraction/osaSocket.h>
@@ -905,7 +906,7 @@ mtsComponent * mtsManagerLocal::CreateComponentDynamicallyJSON(const std::string
 {
     // -1- try to dynamically load the library if specified
     if (!sharedLibrary.empty()) {
-        // create load and path based on LD_LIBRARY_PATH
+        // create load and path based on LD_LIBRARY_PATH (or PATH on Windows)
         osaDynamicLoader loader;
         std::string fullPath;
         // check if the file already exists, i.e. use provided a full path
@@ -913,7 +914,11 @@ mtsComponent * mtsManagerLocal::CreateComponentDynamicallyJSON(const std::string
             fullPath = sharedLibrary;
         } else {
             cmnPath path;
+#if (CISST_OS == CISST_WINDOWS)
+            path.AddFromEnvironment("PATH");
+#else
             path.AddFromEnvironment("LD_LIBRARY_PATH");
+#endif
             fullPath = path.Find(cmnPath::SharedLibrary(sharedLibrary));
             if (fullPath.empty())  {
                 fullPath = sharedLibrary;