+++ cisstOSAbstraction/osaThread.h
@@ -39,7 +39,8 @@ http://www.cisst.org/cisst/license.txt.
 #define SCHED_FIFO 0   /*! No Scheduling Policy available in Windows */
 #endif
 
-#if (CISST_OS == CISST_DARWIN) // SCHED_FIFO is not defined otherwise
+// SCHED_FIFO is not defined otherwise
+#if (CISST_OS == CISST_LINUX) || (CISST_OS == CISST_DARWIN)
 #include <pthread.h>
 #endif